agent/modules/renpho_biometrics/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
structlog>=24.1
pydantic>=2.5
pydantic-settings>=2.1
renpho-api>=0.1
